Change size of left and right display queues

+++ source/main.c
@@ -41,8 +41,8 @@ int main()
 
     // create queues
     display_queue = xQueueCreate(10, sizeof(system_code));
-    left_display_queue = xQueueCreate(1, sizeof(uint8_t));
-    right_display_queue = xQueueCreate(1, sizeof(uint8_t));
+    left_display_queue = xQueueCreate(1, sizeof(int));
+    right_display_queue = xQueueCreate(1, sizeof(int));
     motor_queue = xQueueCreate(1, sizeof(system_code));
     sensor_base_queue = xQueueCreate(1, sizeof(display_setting));
     temperature_queue = xQueueCreate(1, sizeof(int));
